refactor: restructure file folder for select-dropdown
